chore: retryable endpoint validation per package (#389)
